+++ torchmetrics/functional/__init__.py
@@ -45,7 +45,8 @@ from torchmetrics.functional.regression.mean_absolute_percentage_error import ( 
 from torchmetrics.functional.regression.mean_squared_error import mean_squared_error  # noqa: F401
 from torchmetrics.functional.regression.mean_squared_log_error import mean_squared_log_error  # noqa: F401
 from torchmetrics.functional.regression.pearson import pearson_corrcoef  # noqa: F401
-from torchmetrics.functional.regression.r2 import r2_score, r2score  # noqa: F401
+from torchmetrics.functional.regression.r2 import r2_score  # noqa: F401
+from torchmetrics.functional.regression.r2score import r2score  # noqa: F401
 from torchmetrics.functional.regression.spearman import spearman_corrcoef  # noqa: F401
 from torchmetrics.functional.regression.symmetric_mean_absolute_percentage_error import (  # noqa: F401
     symmetric_mean_absolute_percentage_error,